services/hitl_gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot==21.3
aiosmtplib==3.0.1
supabase==2.6.0
google-generativeai==0.8.3
pydantic==2.7.1
jinja2==3.1.4
python-decouple==3.8
structlog==24.2.0
